Use int for temps

+++ WADemo.Core/WeatherRecord.cs
@@ -9,7 +9,7 @@ public class WeatherRecord
 
   public DateTime Date { get; set; }
   public string Description { get; set; }
-  public decimal HighTemp { get; set; }
+  public int HighTemp { get; set; }
   public decimal Humidity { get; set; }
-  public decimal LowTemp { get; set; }
+  public int LowTemp { get; set; }
 }
